refactor: split identity and location modules into single-responsibility files
Extract extension methods from response DTOs into dedicated extension
files and split combined exception handlers into one-handler-per-file.
Move IdentityService to the service package for consistency.
