+++ src/main/kotlin/com/nickdferrara/fitify/identity/internal/advices/UserNotFoundExceptionHandler.kt
@@ -1,0 +1,20 @@
+package com.nickdferrara.fitify.identity.internal.advices
+
+import com.nickdferrara.fitify.identity.internal.controller.AuthController
+import com.nickdferrara.fitify.identity.internal.controller.UserController
+import com.nickdferrara.fitify.identity.internal.dtos.response.ErrorResponse
+import com.nickdferrara.fitify.identity.internal.exception.UserNotFoundException
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [AuthController::class, UserController::class])
+internal class UserNotFoundExceptionHandler {
+
+    @ExceptionHandler(UserNotFoundException::class)
+    fun handleNotFound(ex: UserNotFoundException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+            .body(ErrorResponse(ex.message ?: "User not found"))
+    }
+}
