+++ src/main/kotlin/com/nickdferrara/fitify/identity/internal/dtos/response/RegisterResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.identity.internal.dtos.response
 
-import com.nickdferrara.fitify.identity.internal.entities.User
 import java.util.UUID
 
 internal data class RegisterResponse(
@@ -8,11 +7,4 @@ internal data class RegisterResponse(
     val email: String,
     val firstName: String,
     val lastName: String,
-)
-
-internal fun User.toRegisterResponse() = RegisterResponse(
-    id = id!!,
-    email = email,
-    firstName = firstName,
-    lastName = lastName,
 )
