+++ src/main/kotlin/com/nickdferrara/fitify/identity/internal/extensions/RegisterResponseExtensions.kt
@@ -1,0 +1,11 @@
+package com.nickdferrara.fitify.identity.internal.extensions
+
+import com.nickdferrara.fitify.identity.internal.dtos.response.RegisterResponse
+import com.nickdferrara.fitify.identity.internal.entities.User
+
+internal fun User.toRegisterResponse() = RegisterResponse(
+    id = id!!,
+    email = email,
+    firstName = firstName,
+    lastName = lastName,
+)
