+++ src/main/kotlin/com/nickdferrara/fitify/identity/internal/extensions/UserPreferencesResponseExtensions.kt
@@ -1,0 +1,8 @@
+package com.nickdferrara.fitify.identity.internal.extensions
+
+import com.nickdferrara.fitify.identity.internal.dtos.response.UserPreferencesResponse
+import com.nickdferrara.fitify.identity.internal.entities.User
+
+internal fun User.toPreferencesResponse() = UserPreferencesResponse(
+    theme = themePreference.name,
+)
