+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/advices/LocationValidationExceptionHandler.kt
@@ -2,9 +2,7 @@ package com.nickdferrara.fitify.location.internal.advices
 
 import com.nickdferrara.fitify.location.internal.controller.AdminLocationController
 import com.nickdferrara.fitify.location.internal.controller.LocationController
-import com.nickdferrara.fitify.location.internal.dtos.response.ErrorResponse
 import com.nickdferrara.fitify.location.internal.dtos.response.ValidationErrorResponse
-import com.nickdferrara.fitify.location.internal.service.LocationNotFoundException
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.MethodArgumentNotValidException
@@ -12,13 +10,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler
 import org.springframework.web.bind.annotation.RestControllerAdvice
 
 @RestControllerAdvice(assignableTypes = [AdminLocationController::class, LocationController::class])
-internal class LocationExceptionHandler {
-
-    @ExceptionHandler(LocationNotFoundException::class)
-    fun handleNotFound(ex: LocationNotFoundException): ResponseEntity<ErrorResponse> {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND)
-            .body(ErrorResponse(ex.message ?: "Location not found"))
-    }
+internal class LocationValidationExceptionHandler {
 
     @ExceptionHandler(MethodArgumentNotValidException::class)
     fun handleValidation(ex: MethodArgumentNotValidException): ResponseEntity<ValidationErrorResponse> {
