+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/dtos/response/LocationResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.location.internal.dtos.response
 
-import com.nickdferrara.fitify.location.internal.entities.Location
 import java.time.Instant
 import java.util.UUID
 
@@ -17,19 +16,4 @@ internal data class LocationResponse(
     val active: Boolean,
     val createdAt: Instant,
     val operatingHours: List<OperatingHoursResponse>,
-)
-
-internal fun Location.toResponse() = LocationResponse(
-    id = id!!,
-    name = name,
-    address = address,
-    city = city,
-    state = state,
-    zipCode = zipCode,
-    phone = phone,
-    email = email,
-    timeZone = timeZone,
-    active = active,
-    createdAt = createdAt!!,
-    operatingHours = operatingHours.map { it.toResponse() },
 )
