+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/dtos/response/OperatingHoursResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.location.internal.dtos.response
 
-import com.nickdferrara.fitify.location.internal.entities.LocationOperatingHours
 import java.time.DayOfWeek
 import java.time.LocalTime
 import java.util.UUID
@@ -10,11 +9,4 @@ internal data class OperatingHoursResponse(
     val dayOfWeek: DayOfWeek,
     val openTime: LocalTime,
     val closeTime: LocalTime,
-)
-
-internal fun LocationOperatingHours.toResponse() = OperatingHoursResponse(
-    id = id!!,
-    dayOfWeek = dayOfWeek,
-    openTime = openTime,
-    closeTime = closeTime,
 )
