+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/extensions/OperatingHoursResponseExtensions.kt
@@ -1,0 +1,11 @@
+package com.nickdferrara.fitify.location.internal.extensions
+
+import com.nickdferrara.fitify.location.internal.dtos.response.OperatingHoursResponse
+import com.nickdferrara.fitify.location.internal.entities.LocationOperatingHours
+
+internal fun LocationOperatingHours.toResponse() = OperatingHoursResponse(
+    id = id!!,
+    dayOfWeek = dayOfWeek,
+    openTime = openTime,
+    closeTime = closeTime,
+)
