+++ src/main/kotlin/com/nickdferrara/fitify/notification/internal/dtos/response/DeviceTokenResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.notification.internal.dtos.response
 
-import com.nickdferrara.fitify.notification.internal.entities.DeviceToken
 import java.time.Instant
 import java.util.UUID
 
@@ -9,11 +8,4 @@ internal data class DeviceTokenResponse(
     val userId: UUID,
     val deviceType: String,
     val lastUsedAt: Instant?,
-)
-
-internal fun DeviceToken.toResponse() = DeviceTokenResponse(
-    id = id!!,
-    userId = userId,
-    deviceType = deviceType,
-    lastUsedAt = lastUsedAt,
 )
