+++ src/main/kotlin/com/nickdferrara/fitify/notification/internal/dtos/response/NotificationLogResponse.kt
@@ -1,7 +1,6 @@
 package com.nickdferrara.fitify.notification.internal.dtos.response
 
 import com.nickdferrara.fitify.notification.internal.entities.NotificationChannel
-import com.nickdferrara.fitify.notification.internal.entities.NotificationLog
 import com.nickdferrara.fitify.notification.internal.entities.NotificationStatus
 import java.time.Instant
 import java.util.UUID
@@ -13,13 +12,4 @@ internal data class NotificationLogResponse(
     val eventType: String,
     val status: NotificationStatus,
     val sentAt: Instant,
-)
-
-internal fun NotificationLog.toResponse() = NotificationLogResponse(
-    id = id!!,
-    userId = userId,
-    channel = channel,
-    eventType = eventType,
-    status = status,
-    sentAt = sentAt!!,
 )
