+++ src/main/kotlin/com/nickdferrara/fitify/notification/internal/extensions/DeviceTokenResponseExtensions.kt
@@ -1,0 +1,11 @@
+package com.nickdferrara.fitify.notification.internal.extensions
+
+import com.nickdferrara.fitify.notification.internal.dtos.response.DeviceTokenResponse
+import com.nickdferrara.fitify.notification.internal.entities.DeviceToken
+
+internal fun DeviceToken.toResponse() = DeviceTokenResponse(
+    id = id!!,
+    userId = userId,
+    deviceType = deviceType,
+    lastUsedAt = lastUsedAt,
+)
