+++ src/main/kotlin/com/nickdferrara/fitify/subscription/internal/dtos/response/SubscriptionPlanResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.subscription.internal.dtos.response
 
-import com.nickdferrara.fitify.subscription.internal.entities.SubscriptionPlan
 import java.math.BigDecimal
 import java.util.UUID
 
@@ -9,11 +8,4 @@ internal data class SubscriptionPlanResponse(
     val name: String,
     val planType: String,
     val basePrice: BigDecimal,
-)
-
-internal fun SubscriptionPlan.toResponse() = SubscriptionPlanResponse(
-    id = id!!,
-    name = name,
-    planType = planType.name,
-    basePrice = basePrice,
 )
