+++ src/main/kotlin/com/nickdferrara/fitify/subscription/internal/dtos/response/SubscriptionResponse.kt
@@ -1,6 +1,5 @@
 package com.nickdferrara.fitify.subscription.internal.dtos.response
 
-import com.nickdferrara.fitify.subscription.internal.entities.Subscription
 import java.time.Instant
 import java.util.UUID
 
@@ -13,15 +12,4 @@ internal data class SubscriptionResponse(
     val currentPeriodEnd: Instant?,
     val expiresAt: Instant?,
     val createdAt: Instant?,
-)
-
-internal fun Subscription.toResponse() = SubscriptionResponse(
-    id = id!!,
-    userId = userId,
-    planType = planType.name,
-    status = status.name,
-    currentPeriodStart = currentPeriodStart,
-    currentPeriodEnd = currentPeriodEnd,
-    expiresAt = expiresAt,
-    createdAt = createdAt,
 )
