+++ src/main/kotlin/com/nickdferrara/fitify/subscription/internal/extensions/SubscriptionPlanResponseExtensions.kt
@@ -1,0 +1,11 @@
+package com.nickdferrara.fitify.subscription.internal.extensions
+
+import com.nickdferrara.fitify.subscription.internal.dtos.response.SubscriptionPlanResponse
+import com.nickdferrara.fitify.subscription.internal.entities.SubscriptionPlan
+
+internal fun SubscriptionPlan.toResponse() = SubscriptionPlanResponse(
+    id = id!!,
+    name = name,
+    planType = planType.name,
+    basePrice = basePrice,
+)
